fix: Remove resource_name from monitoring codes
