--- src/components/Modals/Monitoring/Multiple/index.scss
@@ -1,2 +1,0 @@
-@import '~scss/variables';
-@import '~scss/mixins';
