--- src/pages/clusters/components/Cards/Monitoring/UsageRank/index.scss
@@ -1,63 +1,0 @@
-@import '~scss/variables';
-
-.wrapper {
-  border-radius: $border-radius;
-  box-shadow: 0 4px 8px 0 rgba(36, 46, 66, 0.06);
-  background-color: #ffffff;
-  overflow: hidden;
-}
-
-.inner {
-  border-radius: 4px;
-  padding: 12px;
-  background-color: $lightest;
-}
-
-.tabs {
-  cursor: pointer;
-  padding: 2px;
-  border-radius: 18px;
-  display: inline-block;
-  background-color: $bg-color;
-  border: solid 1px #d8dee5;
-  & > span {
-    display: inline-block;
-    line-height: 32px;
-    color: #79879c;
-    padding: 0 30px;
-  }
-  @at-root {
-    span.selected {
-      border-radius: 16px;
-      background-color: #36435c;
-      color: #fff;
-    }
-  }
-}
-
-.selectContainer {
-  float: right;
-  margin-top: -55px;
-
-  :global .select {
-    width: 250px;
-
-    & + .select {
-      margin-left: 12px;
-    }
-  }
-}
-
-.link {
-  line-height: 32px;
-  border-radius: 4px;
-  background-color: $bg-color;
-  text-align: center;
-  a {
-    color: #329dce;
-    display: block;
-    &:hover {
-      text-decoration: underline;
-    }
-  }
-}
