+++ src/pages/clusters/components/Cards/Monitoring/index.js
@@ -20,4 +20,3 @@ export { default as ClusterResourceStatus } from './ClusterResource'
 export { default as ServiceComponentStatus } from './ServiceComponent'
 export { default as ETCDStatus } from './ETCD'
 export { default as ETCDNodes } from './ETCDNodes'
-export { default as UsageRank } from './UsageRank'
