+++ src/pages/fedprojects/containers/Base/Detail/index.jsx
@@ -20,8 +20,8 @@ import { inject } from 'mobx-react'
 import { withRouter } from 'react-router-dom'
 import { Component as Base } from 'core/containers/Base/Detail/Page'
 
-@inject('rootStore', 'projectStore')
 @withRouter
+@inject('rootStore', 'projectStore')
 export default class DetailPage extends Base {
   get enabledActions() {
     const { cluster, namespace } = this.props.match.params
