+++ src/pages/projects/components/Modals/PodsMonitoring/index.scss
