+++ src/stores/monitoring/container.js
@@ -19,6 +19,8 @@
 import Base from './base'
 
 export default class ContainerMonitoring extends Base {
+  resourceName = 'container'
+
   getApi = ({ namespace, podName, container }) =>
     `${
       this.apiVersion
