+++ src/stores/monitoring/node.js
@@ -19,5 +19,7 @@
 import Base from './base'
 
 export default class NodeMonitoring extends Base {
+  resourceName = 'node'
+
   getApi = () => `${this.apiVersion}/nodes`
 }
